refactor: 💡 move function for handleWin
